clean up fn defns and imports
